requirements.txt
kafka-python==2.0.2
websockets==13.1
